upd commands partly
commands was partly updated for new api and moved to Single Folder
