+++ Commands/Single/GetFollowersCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetFollowersCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_followers';
+}
